Fixed, termination Bug(s)!
This was pretty nasty. The hardest thing to track was that I accidentially terminated states BEFORE removing them from the mapper.
Note that this is AGAIN an instance that was caught by our SafeList! (just saying)

Another bug that was related to termination was that in the single-test-case mode, not all dscenario memebers were reported to KLEE for termination (hence only one state was removed from the searcher, while the other would go on forever).