CoojaInformation intialisation bug
The problem was a sloppy typo in the Copy Ctor of CoojaInformation.
The danglingTimes were not initialised to the original's scheduledTimes, but to the own scheduledTimes which was necessarily empty on construction. This caused the adding procedure to schedule the state to the default time, ignoring the events the user-code thought were scheduled.