FIXING SCHEDULING BUG
This is a big one, even if the changes seem small!
The FES was running empty for the reason that there were unclear semantics about scheduleState. At some point it was understood as abolute and at some point as relative, causing the times to vary between huge and tiny.

This was fixed by completely removing scheduleState and replacing it with
  scheduleStateIn and
  scheuldeStateAt
This removes the ambiguity and allows users to still use both semantics.

Now, the old tests run through!