fixing termination bug
This wasn't really as much as a bug, as more a case of overzealous asserting.
StateMapper::terminateCluster is invoked by KleeNet for any state that dies before its time, even on those that were already part of a cluster that was terminated (and have no MI anymore). Our code fetched the MI and tested whether it was still there. However, it used a convienience function (stateInfo) that already asserted that NULL is never returned.