fixing issue #503 (terminateState from user-code)
Calling terminateState from silent_exit (from model code) works now as intended because we implemented terminateState in the NetExecutor, overriding (EXTENDING) the implementation in the klee Executor.
Note that this adds some overhead when terminating states, because we have to cut through a couple functions more, but that shouldn't concern us right now.