implementing/fixing mergeConstraints
The new implementation of this is now only used when resolving transmissions with the TransmissionHandler.
We first grab all expressions and use the klee ExprVisitor to exchange all Read expressions with expressions that read from a cloned symbolic object.

So in an expression (think of expressions as a tree) that is to be sent via the virtual wire, we could have tree-nodes that read from symbolic locations.
E.g. Reading index 0 from "foo" on a state that represents node 42, being sent to node 74.
Since this will travel to completely different ExecutionState(s), which may in turn have its/their own symbol called "foo",
we have to do something with this object, before accidentally squashing different things together.
So we first create (on demand) a new klee::Array object and try to find a unique name starting with "foo{42->74}".
If that name is already used (perhaps due to prior transmissions), we try "foo{42->74}(2)", then "foo{42->74}(3)", and so forth.

This lazy translation is applied to all expressions that are written into the target's "dest" variable as well as all constraints of the target state.
Note that we just copy **ALL** friggin constraints, no matter what.
This is excruciatingly inefficient and dramatically blows up the constraint set of the target state.
However, it is pretty hard to determine which constraints we need and which we don't -- in an efficient way.
This should be fixed in the future.