made SafeList work with other things than pointers
So far the SafeList would refuse to compile or produce wrong results if used with anything other than pointers.
While keeping track of pointers is it's main purpose (that's why it has to be devilishly fast, which it isn't as much as I'd like), it should at least be possible to use it with more complicated aggregates, without having it create default instances all other the place.

Note that the implementation is a lot less efficient for primitive types other than `pointers` and `int`s but proper specialisations can be made!