fixing SpecialFunctionHandler to use klee_message
Our implementation of SpecialFunctionHandler was dumping all calls if the user requires (this is not an internal debug dump, but to be kept when shipping).
This was bogus, and changed to use klee_message instead of std::cout. Formatting is done with std::ostringstream.
This is important, because this also ends up the messages.txt file and is prefixed with 'KLEE'.