"single-container-mention-iterator"
Sometimes containers are built on-the-fly, so you want to only write them down once, which is usually
not that easy with the C++ iterator idiom:
for (Container::const_iterator it = myContainer.begin(), end = myContainer.end(); it != end; ++it) {}

If we only want to name myContainer once, or we simply want to use a different iteration paradigm, we can use
for (LoopConstIterator<Container::const_iterator> it(myContainer); it.more(); it.next()) {}

This will allow us a couple extra features like counting, etc. And it's often shorter.