+Dictionary Meta Container
Using indirection we can abstract away the difference between std::vector and std::map and regard them as the same thing. This simple container (Dictionary) allows us to query for indices and for keys, which coincide when a continuous dictionary is requested, as we use a simple vector in that case. Otherwise we use a map and a vector: The map as key to index translator and the vector to translate from index to key and from index to value.

This allows us flexible and efficient implementation of the BinaryGraph for constraints-fiddling.