refactored the BinaryGraph
I'd say the BinaryGraph is complete now. It can efficently search for basically everything (maybe we can use it later to enhance our SdsClusterer.

However, it will use two net::util::Dictionary datastructures to represent both partitions of the verteces (i.e. constraint index and Array const*). It works entirely on indices when performing the search, so we don't have to lookup tree values from a map all the time (only when constructing and back-translating, we have to use the map).

I shall run some performance tests when I get to implement the TransmitHandler, which currently doesn't call the BG search, yet.