adding a tx counter to BasicState
Although we do have our fancy dynamic runtime plugin, this is something that we expect to always need, and it's a small thing to carry around. Note that the State type that is used when compiling KLEE (as opposed to KleeNet) will *NOT* aggregate the net::BasicState, so it doesn't cost anything there. And in KleeNet we do need this statically for resolving and managing transmission data.