kleenet::State persistent transmission helper-data
While still keeping the kleenet::State as small as possible (so it doesn't clutter KLEE), we now have a possibility to add persistent data that helps us speed up transmissions (a lot). We use the compiler fire wall to make sure nobody does evil stuff with that pointer (nobody has a clue what it is). Unfortunately, we cannot use a smart pointer here, but have to resort to dynamic binding because we want the state to get rid of that information itself, when it dies (delete needs a complete type). Note the virtual dtor!