more sensible index vs. key semantics
The net::util::Container is specialised datastructure to handle dictionaries.
However as opposed to a map (which has a key and a value) and a vector (which has an index and a value) our container has a key, an index and a value.
Access by index is always fast (amort. constant) while access by key depends on the traits of the key: if the key is integral (something equivalent to size_t) and CONTINUOUS (i.e. starts at 0) then lookup by key is also fast.
Otherwise it's logarithmic (as the default map implementation is used).
Regardless an index is always provided for fast lookup (the map may merely translate a key to an INDEX, not a value).

An index works like a relaxed iterator (especially an untyped iterator!).