Total (incomplete) overhaul of transmit handler
<important>

The transmit handler design we had so far was okay at first glance, but prohitively wasting computing ressources at closer look.
For each transmission (deliveries to targets are considered SEPARATE transmissions!) it would build all datastructures and translation matrices FROM SCRATCH.
While this already worked very basically, it was crap in the long run.

So now everything is thrown into a persistent kleenet::State member, that allows us to pick up our work where we left.
This means that transmissions of the same packet to separate targets will effectively result in almost no work because we cached all relevant data before.
Also, new packets will be processed faster, because the long-lived data-structures (like the constraints-array bipartite graph) are preserved and only updated if needed.

Note that this overhaul effectively removed some already existing features, which are to be added in future commits.
The existing commit is already monolithically huge, but not divisible into smaller chunks.