add nodes by iterators to Bipartite Graph
The classical container "add stuff" idiom is: give me two iterators, begin and end, and I will add everyhing in [begin,end).
So far we took a templated container and determined its begin and end position ourselves.
This is cumbersome, and doesn't allow us the flexibility of that other idiom, as we cannot choose where to start outside of the container.