changing the bipart-graph node types in tx hnd
Logically, we need to traverse the graph that consists of constraints (node type 1) and read expressions (node type 2).
We have been implementing this by using a bipartite graph that stores as node type 1 a `vector::size_type` (which represents the index to the constraints_ty vector in the ContraintsManager) and an as node type 2 and `Array const*` that represents the read expression.
This is a bit cumbersome, because after the biparite graph is done we have translate the vector back to actual expressions by looking the indices up in the ConstraintsManager.
However, the Containers used in the BGraph actually handle arbitrary types by indexing and thus already offers us that exact feature.
Advantage: we don't have to use the awkward translation code and can benefit further from rvo.