Symbol propagation
Now the new symbols are correctly written to the target state after mangling.
The symbolic objects (klee::Array) are connected to an ad-hoc fresh MemoryObject to be displayed correctly when analysing the ktest file.
This is a bit tricky and might still be wrong, because we pass NULL for the llvm::Value.