add return value to StateMapper::terminateCluster
To better make sense of the number of clusters the StateMapper routine that takes care of killing an entire cluster returns if the pivotal state (the one to figure out WHICH cluster to terminate) actually referred to a cluster / was known to the mapper. This may not be multiple-state-mapper-safe (but we're not using mutliple mappers at the moment).

StateMapper::terminateCluster will return true iff the passed state had a valid cluster attached and that cluster was removed.