partly fixed cluster counting bug
Previously clusters were often counted twice due to compliated call path during state termination (I don't claim it's perfect, but it does have some consistency).
Anyway, now we're using the information prodived by StateMapper::terminateCluster to correctly determine whether to count a particular cluster.

Note that this still produces a total number of clusters that is too large, but it is (presumably) the correct number of clusters that are created during the execution, because while splitting up clusters, new clusters are created.