flexible debug infrastructure
Okay, here's the deal: These pesky std::cout's just didn't cut it anymore.
The vast flood of debug messages from various modules hinder readability.
The problem is, if we're debugging something, and add status messages via cout/cerr, and are "done" with that feature, we have to remove them again, so they won't clutter everything.
But we might want to go back to debugging that thing, or the debug statemens were helpful, so we may only comment them out.

This back and forth generates NOISE in the git log! Noise is bad!

So, now we have a debug infrastructure in the net module that can even be lent by, say, the kleenet module.
The deal is that we include "net/util/debug.h" and write stuff to e.g. net::DEBUG<net::debug::mapping>::cout which then decides if it's discarded or actually written to cout.
This can be trivially aliased via the proprocessor INSIDE A CPP FILE (never do this in H files) to something like
  \#define net::DEBUG<net::debug::mapping> DD
which allows us to write
  DD::cout << "hello" << DD::endl;
while the debug.h decides which "channels" are selected to go to the out buffer. For instance we could say we want
  ENABLE_DEBUG (mapping | searchers | external1)
such that stuff written to DEBUG<mapping> would be printed but something written to DEBUG<slack> wont.

** Note that the defines/macros in debug.h are ALL undefined before the end of debug.h **