CHANGING KLEE CODE
To implement distributed constraints, we need to have notion of Arrays that belong to each other.
We can extend klee::Array to do this nicely, but whoever gets rid of the klee::Array objects, will only have a klee::Array* handle, so we need to add a vtable to klee::Array, unfortunately.

Upstream will *never* accept this, I'm afraid.
If we really want to up-merge some day, we may move this to a map lookup or some other ugly business.
However, there are more problematic things than this, so unless, upstream offers a better interface for extensions, we have little choice to extend KLEE, other than these small intrusions.