Infrastructure to track distr. constr.
Distributed Constraints is a conceptual design which allows several states to have symbolic objects that logically belong to each other.
To implement this, we extend the klee::Array with a DistributedArray which is invisible outside the DistributedConstraints.o compilation unit.
It behaves just like a klee::Array.

However, it allows us to add bind several Arrays together and allow (re-)transmissions of symbols to be identified correctly.

We have yet to add this facility to the TransmitHandler transmission handling.
While the state representative has already been added to the ConfigurationData, it isn't used by the LazySymbolTranslator, yet.