transmission destination bug fixed
There was an error in handling the destination node id for transmissions.
While an ConfigurationData object is only constructed once for a given state, a TxData object is constructed for each invidiual transmission (wich a distinct tx id).
However, different transmissions can have different destination node, thus the destination node id must be a property of the volatile TxData, not the persistent ConfigurationData object.