separating per-receiver and general tx data
This is an important distinction that we didn't do previously, and that could be characterised as a bug.
Actually I would have a hard time, arguing how this would have been okay, even without phony transmissions.

The thing is that if we have a transmission from a state s to a target node i, then there may be several (say k) receiver states r_1, ..., r_k.
If we only consider a TxData context, which is instantiated for each new unique transmission-id (a transmission-id characterises individual transmissions, even in respect to phony packets), then all individual deliveries (s,r_j) with 1<=j<=k would be delivered using the same caching data structures.
Hence, the translation-matrix, which is in itself designed to be lazy (this isn't a simple runtime improvement, it affects *correctness*, because if it wasn't lazy, it would re-translate the same Array object over and over and over, which would result in different objects, where only one was intended) will not understand that now a different receiver state is used and simply use its old cache, assigning Array that were intended for r_1 to r_2, ..., r_k.
There is data that is okay to be used for different r_j, such as the *sender* symbols, as they *do not* change, we only need those to extract the necessary constraints later on.
However, the receiver specific objects *must* be purged after one receiver state is done.
These include the LazySymbolTranslator and its associated NameMangler (which is about to get more complicated in future patches, and depend heavily on the specific receiver *state*, not only its node id (as it does currently)).

Now, we instantiate a PerReceiverData *on the Stack* of TransmitHandler::handleTransmission as it will only be needed in that very context:
Once that function is done, that particular receiver state will not be touched again *for this transmission-id*.

Note that the ReadTransformator, which is also receiver-specific was merged with the ReplaceReadVisitor type, as the latter is never needed otherwise, and we can save some extra pointers and enhance readability (mainly due to abstraction).