caching required sender constraints separately
It seems we were a bit too quick: The required constraints is something that we *can* and *should* cache, as they won't change.
However, applying the ReadTransformator operation is something we have to do individually for every receiver state, so we use two separate vectors, to track the respective expressions.