implemented the Aliasing NameMangler
The basic functionality of the Aliasing NameMangler works.
Note that we will have to add additional constraints, for it to produce less false positives (for it to produce *none* at all, we will have to interface with the mapper).
It can be selected via a command line switch: -distributed-symbol-name-mangling={duplicating,aliasing} while 'aliasing' is default and 'duplicating' is the old(er) working mangler, that translates symbols all the time.