computing sender Arrays to make EqExpr of
In the new NameMangler scheme, we have to add extra EqExpr for the sender if a fresh non-distributed symbol is distributed.
This means that we have to caputre the transition from non-distributed symbol (e.g. "foo") to distributed symbol (e.g. "foo{node1:tx1}") and construct an equality constraint to make sure these values are always the same.
Adding the equal constraint is yet to be implemented. This commit contains the computation of the actual Arrays that are to be equalled.