fixing DistributedArray rtti guessing game
Determining if a klee::Array* is in fact a DistributedArray* is a guessing game and sits there just begging to break.
While the current solution *could* break as well, if somebody else were to extend klee::Array and use our facility to say that something is not a klee::Array, this scenario is not
We're the only ones who have access to this method, as I added it for KleeNet.