sane way of handling new symbols during tx
Now, all (potentially) new symbols are collected and sorted into the respective state, regardless of *which* state the are supposed to belong to.
Also, we have some additionally deprected stuff now, that we have to get rid of ASAP, as it does redundant work and/or is dangerously ill-defined and may end up breaking transmission code.