added --omit-size flat to ktest-tool
Who'd believe it, I messed with python soup ... anyhow, it's a bit distracting to see the 4 pop up all over the place when quickly parsing the result of `ktest-tool --write-ints`, most of the time I just don't care about that.

It's gonna be fun patching that thing to make it understand SDE specific stuff, because right now, it doesn't have a clue what node/dscenaio a ktest is about.