allow and use symbol precomutation
When preparing messages we have to precache the translations.
This means that the main translation work is done in the ctor of the PerReceiverData constructor which holds the context for the current stack frame.

This allows us to simply query properties like size of the table and so forth without having to worry that additional symbols will be added in the future.