removing defunct MO-per-symbol code
Previously we thought about adding a new MemoryObject for each individual Array object that is transferred to the receiver state.
This doesn't work and produces the wrong result as we don't actually want testcases for each object that *includences* what gets written.
We care about what actually travels through the wire.
This particular thing is added later on, but the current patch removes the wrong implementation of the old approach.