main computation of "wire-content"
We can only singlecast, state-wise.
Each individual transmission can be modelled as a data packet being delivered via virtual wire to the receiver state (note the distinction between "receiver state" and "target node").
What arrives in the input-buffer of the receiver state is what we consider the truly delivered packet (which might of course be symbolic, concrete or mixed).
For each transmission that is not entirely concrete, we construct a new symbol (which actually includes a new MemoryObject that is bound to the receiver (and only that) via a new ObjectState) to be created.

Now, the actual input-buffer goes ahead and simply reads (byte by byte) from the newly created object (first line in the inner loop).
So far, the array is naked and we haven't put our actual packet data anywhere.
To make use of our packet data, we add individual constraints that bind each byte of the input-buffer to a byte in the newly created array.

This will allow the solver to actually generate the desired values for the packet object (which has the format `tx<<tx-id>>(node<<node-id>>)`, for example tx74(node12) for the 74 transmission made by node 12).