fixing Bug #561 LockstepSearcher access NULL
When consolidating its cache, it fastforwards null entries, which is perfectly fine.
However, if it is at it's end, it rewinds its iterators, so that the selectState routine can go on ad infinitum.
Bug if the first entry in its cache happens to be a removed state, it will be 0, in which case it was actually operated on by selectState.
By adding a simple additional fastForwardJunk call after resetting the iterators everything's fine, as we catch this.