fixing Bug #562 inf. loop in lockstep searcher
If a blocked state was transmitted a packet, its blocked flag was correctly copy-constructed, but the internal caching that speeds up queries that test if all states are blocked (to unblock them) was not updated.

Now, when adding a state we simply check for its blocked flag to be set and, if so, simply increment the counter that tracks the number of blocked states known to this particular searcher.
Note that the ::remove method already correctly took care of this.