adding a type for node-id transformation
The user is obliged to using ids >= 1.
More to the point we allow them to use any kind of ids, for instances node 1, node 2 and node 7, but none else.
This means wherever our internal stuff stores something it's either expensive lookup (std::map) or expensive storage (std::vector / std::array / c-array).
So the idea is to eventually move to a point, where the user can use *any* collection of node ids (for instance, you could imagine using ip addresses!), while in the engine we would be using continuous ids, staring from 1 and having no gaps to speed up everything.

This commit only adds the type that I propose to deal with this, however, I haven't applied this, because if I do, everything the engine leaks out NOT via some special function handlers will use the internal ids, which is problematic for
  a) error messages
  b) testcase output
Before we have a good idea about those, the transformation will not be applied.