fixing a terrible bug in LockStep!
I'm at a loss how this went unnoticed for so long. It could explain the weird behaviour we saw on a different platform/compiler.
What I completely ignored here is that vectors a jumpy little beings: if they run out of *capacity*, they move their payload somewhere else.
If you're "lucky", a small realloc will not move the memory location, but there's always the possibilty that a push_back will do this.
This is the reason, iterators are invalidated on mutating operations.
However, we kept track of the currently selected state via the `next` member, which was a vector's iterator.
If we push, it points to old (i.e. deleted) memory, and what we read is garbage (possible segfault).

This is fixed by completely recomputing the iterators on insertions in ::add.
When we are consolidating the cache we already took care of our iterators (because it was more obvious that they would be gone there!).

Now they are handled correctly on all mutations.