+ DynamicFuntor, and FunctorBuilder
To decouple modules it is sometimes useful to pass functors (used as lambdas with small captures, in a sense) around, whithout fully typing them from the caller side.
The dynamic functor allows us to specify that we expect *something* that can handle an object of a given type.
The derived Functor<T,F,DynamicFunctor> then fills in that functionality with a concrete F, which does the actual work.
Note that we can also use a Functor<T,F,StaticFunctor> (which is the default Functor<T,F>) if we do not require dynamic binding!