fixing bug #572 send+fork+send segfaults
The problem was the wrong assumpting that the ConstraintSet only grows.
However the ConstraintManager can remove AND change constraints.
In addition the loop that used an integer as iterator did an `it != end` comparison instead of `it < end` which had the result that as the number of known constraints dropped from 4 to 3, the loop would start at position 4 (as it assumed 0..3 were already processed) which clearly is not equal to 3, so the loop went on to read past the end of a vector causing mayham.

In a way this bug was a good thing, because it hinted us at a far greater bug.
If the condition *would* have been `it < end` the error would have been silently dropped and new constraints that were placed in an *old* slot (wrongly presumed to be already processed!) would have gone unprocessed.