issue #496 -- transferConstraints done
Rejoice! It is finally implemented.
With the massive data structures for transmitting Packets (used in and designed for the transmission handler) we could relatively concisely (if you don't regard all the debug messages I still left in there) implement this.
We simply extract all DistributedArrays we have local instances of, give that thing to the BipartiteGraph and it kindly returns all Constraints (and possibly completely new Arrays) we have to transmit. Magic.

The debugging output is terribly verbose, but it helps to figure out what is going on, while we are still in a ... sanity checking phase.
Of course it should be reduced eventually, even if dormant (optimised out) if you set the right flag in include/net/util/debug.h.