overriding caching options in KleeNet (#573)
Caching of constraints can lead to unexpected (for the KLEE part) situation as the KleeNet code messes with the internal data structures of KLEE, adding constraints.
This can lead to unsatisfiable states, as we cannot ensure the complete absence of false positives (at the moment).

With this example of a totally and utterly over-engineered solution for an otherwise trivial problem (sorry for that, couldn't resist), we override the default KLEE setting for the cache and cex-cache (-use-cex-cache and -use-cache) options to be false, while allowing the user to prevent this overriding by issuing -sde-override-cache=0 or -sde-override-cex-cache=0 which both default to 1 (true) and have the semantic that KleeNet is allowed to set the default value of true to false.

If the user knows what they're doing, they can still have the default caching or cex-caching (or both), at their own risk.
But if not explicitly stated, KleeNet will force both to be off (even if the value is explicitly stated on the command line).

This is an important compatibility issue to be able to compile klee (!) with the same sources (and .o files!) as kleenet;
If you compile klee you expect the same defaults (and results) as if you would compile klee from the KLEE-repository.