add transformating std iterator to net module
So far we already had an StdConstIterator to use our ConstIteratable interface with std containers.
Now we add some functional awesomeness and allow to use arbitrary transformations.

This means that we had to change the return type from ConstRef<T> to RemoveRef<T> to avoid returning temporaries from the transformation function ("lambda").