adding different transmit modes and flexible input
The "tx13" or "pull13" is now computed earlier in the stack to allow more flexibility, which means that we pass std::string objects around instead of size_t indices.
This could even be extended, but at some point we would need better means that to throw everything into the BasicState.