moving constraint magic to dedicated type(s)
The ConstraintSet type hides most of the ugly caching and weird interfaces that we need to handle constraints/transmissions and so forth.
It will automatically add new symbols to the appropirate states and can be used easily to extract constraints from a given set of expressions OR to extract all constraints that depend on distributed arrays.
This is pretty handy, as we can write stuff like:

  vector<RefExpr> rangeConstr = ConstraintSet(TransmissionKind::tx,sender,expressions.begin(),expressions.end()).extractFor(receiver).extractConstraints();
  // or
  vector<RefExpr> allConstr = ConstraintSet(TransmissionKind::tx,sender).extractFor(receiver).extractConstraints();

This instantiates all persitent and semi-persistent objects along the way and constructs temporary constraint sets (first a ConstraintSet and then a ConstraintSetTransfer object).

Note how much this shrinks the handleTransmission and the transferConstraints methods (it will do the same to kleenet_pull/get_symbol).