fixing critical Lockstep bug
When rebuilding the cache after only a fourth of the size is actually used, the Lockstep searcher didn't update the slot information of the states.
That's pretty bad, because once the vector is shrunk, the states point to the slots in the old vector, clearly crashing (or due to my careful nature SIGABRTing).