working kleenet_reverse_memcpy implementation
This is the core of what kleenet_get_global_symbol does as well.
It is basically the same, only that we are referencing the source/dest objects differently.

The basic idea is that we reverse information flow because this allows us to avoid mapping (if you think about it, it's an implicit packet cache application, because all rivalled states are "sending", but we can even allow them to "send" *ANY* data, it doesn't have to be concrete, it doesn't even have to be compatible, let alone the same).

So a state sucks in all values of its targets (its targets are precisely the states that have this state as target, as visibility is symmetric).
If one state has the value var = 3 and another has the value var = 7 then the pulling state (i.e. the one calling kleenet_reverse_memcpy) will have a variable that evaluates to var in {3,7}.
This is done with constraints trickery: Assume we have states s_1 ... s_k as data sources.
Let's call the expression of state s_i's variable var just x_i and its constraints (i.e. the transitive closure or all constraints that either influence var or depend on distributed arrays) as y_i.
We then construct the mega-constraint:

  OR_{i in {1,...,k}} ( (v == x_i) and y_i )

Where v is a completely new array, which is written byte by byte to the variable of the calling state (the one where the data is flowing to).

The rationale is this: The solver will have to make at least one of theses conjunctions (parts of the big disjunction) true in order to satisfy the mega-constraint.
v therefore represents all values that equal one of the x_i while at the same time satisfying the corresponding y_i.

Note that the entire transmission only adds a single constraint.
Not by choice but by necessity: I don't see a way to represent this as several (smaller) constraints.
That thing is huge; Especially since there are a lot of trivial layers in there.
Stuff like (And true ...) or (Not (Eq false (Eq true ...))) really don't help, but are a byproduct of the way we build the expressions.
I have the dim hope that the constraint manager will simply strip them away, before the solver even sees them.