GlobalValue name caching for faster MO retrieval
When trying to get our hands on a MO for a GlobalValue, which we only know the name of, we have to search the entire collection of GlobalValues until we find the one we are looking for.
If there n GlobalValues and the average name length is m, this takes Theta(n*m) in the average.

Now, we cache all names in our own map, such that we get an amortised O(log(n)*m).

If we used a better datastructure than an std::map (e.g. a Trie), we could even get away with O(m).
But this is really not that important, logarithmic is satisfactory for the moment.