acquireExprRange overloads
We need to specify the location of data in different ways, so we need acquireExprRange to have several overloads that can take different kinds of location information.
One is just giving it the address in form of an ObjectState and an offset, this is the most basic form which the other forms are wrappers for.
Another way is to pass it a MemoryObject, which implies an offset of 0. This isn't theoretically the most general way, but we don't need to add an offset at the moment.
And the oldest form is to pass a ref<Expr> which then is used to determine MO, OS and offset.