StateClusters delete themselves once their empty
This inconsistency led to the cob breaking and SDS leaking, because it's hard to decide when and by whom a StateCluster is to be deleted.
Now it does it itself, and thanks to the Observer everybody who is storing handles to it will be notified.

Sometimes I have to admit ... Patterns *are* handy. Sometimes.