fixing false positives silent default term handler
State::transferConstraints already told us whether it was successful in transferring the constraints from state to state.
The termination handler now acknowledges this information by using a different termination handler in case of unfeasible dscenarios (!).

Note that this is still kinda bullcrap, because we still generate those pesky false positive dscenarios, but simply don't generate test cases for them and don't increment the total-dscenarios-counter.
But unless the mapper becomes smarter when exploding scenarios, we're stuck with that.