terminate on Error vs Early patch
If a state terminates on error, only states that are descendants of this state (due to explosion, e.g.) will generate error files.
This is accomplished by flagging that state with an ERROR flag before handing it to the terminateCluster method.
The flag will be inherited, and we can later check whether a state has its ERROR flag set, when trying to figure out how to handle the final termination request coming from the terminateCluster wibbly-wobbly.