fix issue #627 (duplicate test cases)
The problem was that when forcibly terminating everything, KLEE went through all its known states (Executor::states) and terminated each one of them.
However, in the mean time the KleeNet part would intercept this and not only terminate one state but an entire dscenario.
This led to some states being terminated more than once, when the KLEE part iterates over a state that was already terminated previously.

The fix now does this: We check whether a given state is still known in KLEE (regardless whether KleeNet knows it) and if it is we proceed in terminating it.
Otherwise we silently ignore the termination request.

Although this is not the best solution it is not as hack-ish at it may sound.
One could say this is actually a bug in KLEE as it doesn't check whether a state was already removed.
However, this problem would never occur in KLEE, as states are stictly independent.
Terminating a state couldn't possibly influence other states, so nobody thought of sanity-checking termination requests.