fixing "unknown states bug" in cooja-searcher #629
The problem was that states could have been scheduled (specifically with wakeup_dest_states) before being properly added to the Cooja searcher.
The attached scheduling information was correctly instantiated and everything, the searcher just doesn't like getting a state that was already scheduled added.
Also, it is dangerous to have a state in the calqueue that wasn't added, yet.

The fix is to track the "is added" status inside the scheduling information.
If a state that was *NOT* properly added is scheduled, it is merely added to the "danglingTimes" set.
Only when the state is properly introduced to the searcher (via +=) the dangling events are transformed to proper events.