fixing phonyMap bug #631 "empty dstate node-slot"
When populating the new dstate, the phonyMap altrithm moved every vstate of the sender's node in the old dstate to the new dstate.
This is obviously beyond stupid, because the rivals must remain in the old dstate.
The problem caused severe crashes, as the intersection of a node and a dstate must never ever be empty.

I distinctively remember having fixed this bug before, but I couldn't find the solution in any branch. It might have been reintroduced somehow. It's a mystery.