wakeup_dest recursion in PacketCache
This was a weird one.
When yielding a state the CoojaSearcher (CS) would go and remove that state from its calendar queue (which is perfectly fine).
However, in addition if there are no more events at that same time, the CS would ask the PacketCache (PC) to commit all mappings and deliver the data. Still fine, so far.
But, since we introduced the wakeup functor, the PC would go and execute the post-commit-hooks, including the wakeup call. Still everything fine.
Now, the wakeup would schedule an event at time "right now" for the destination node, which causes the CS to drop future events (remember: CS can have only one event per state, this is a limitation of cooja, which we must mirror), soooo it goes and removes states BEFORE adding the new event. This is still technically okay.
Now it gets messy, when removing the state/event, the calqueue for that time slot is still empty, and the CS goes and asks the PC to commit.
However, the PC was ALREADY (!) right in the middle of committing, so it was just iterating over its packets (stored in an `std::map`).
Innocently enough, the PC iterates over its internal data structure, handles all events (which does not schedule new events, as there is no wakeup hook left), clears its collection of packets and returns.

Is it still fine?
No. The first call to the PC was right in the middle of iterating over a map, and now that map has been cleared of content.
That means all iterators have been invalidated.
From this point only chaos and mayhem can follow as we are accessing garbage data.

The fix is to first swap the internal map into a local buffer (sitting on the stack!) and iterate over THAT.
When the recursive call occurs, the map will already be empty and the commit will (well, MUST in fact) return immediately.
No chance to invalidate iterators.