allowing null-cloners for sparsely used dependants
Kinda negates the original purpose of the StateDependant mechanism, but it is so insanely convienient that we can do all sorts of stuff, so it makes sense to allow a particular SD-type to *not* clone itself to all states.
A null cloner will return null instead of actually cloning an object.