finally fixing the wakeup segfault
Although the PacketCache was correctly rid from removed states, they were still lingering in the WakeupFunctor, in form of dangling pointers to deleted objects.
We already have means to figure out the validity of states in NetExecutor, so we just use them.
This **is** slightly dangerous, as the old pointer *could* have been instantiated as a new state, but that state shouldn't have been able to send data, so it cannot end up as wakeup functor.